\name{info}
\alias{info}
\docType{data}
\title{
Marker Information of The Example Data
}
\description{
The marker information of the example data.
}
\usage{data(info)}
\format{
Column 1: marker ID.
Column 2: chromosome.
Column 3: chromosomal position in base pairs.
}
\details{
The markers are from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Shen X, Ronnegard L and Carlborg O (2011): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. BMC Proceedings, 5(Suppl 3):S14.
}

