\name{pt}
\alias{pt}
\docType{data}
\title{
Phenotypes of The Example Data
}
\description{
The phenotypes of the example data. 
}
\usage{data(pt)}
\format{
Column 1: individual ID.
Column 2: the simulated quantitative trait.
Column 3: the simulated binary trait.
}
\details{
The phenotypes are from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Xia Shen, Lars Ronnegard and Orjan Carlborg (2010): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. (14th QTLMAS, Poznan, Poland) SUBMITTED.
}
