\name{snpid}
\alias{snpid}
\docType{data}
\title{
Indices of The Selected SNPs of The Example Data
}
\description{
The indices of the selected SNPs of the example data.
}
\usage{data(snpid)}
\format{
a numeric vector containing SNP indices.
}
\details{
Every 6 SNPs were selected from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Xia Shen, Lars Ronnegard and Orjan Carlborg (2010): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. (14th QTLMAS, Poznan, Poland) SUBMITTED.
}
