\name{gt}
\alias{gt}
\docType{data}
\title{
Genotypes of The Example Data
}
\description{
The genotypes of the example data with in the coded format. 
}
\usage{data(gt)}
\format{
The matrix is 2326 (individuals) by 1628 (SNPs).
}
\details{
The SNP markers are selected from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Xia Shen, Lars Ronnegard and Orjan Carlborg (2010): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. (14th QTLMAS, Poznan, Poland) SUBMITTED.
}
