\name{pedi}
\alias{pedi}
\docType{data}
\title{
The Pedigree Structure of The Example Data
}
\description{
The pedigree information of the example data.
}
\usage{data(pedi)}
\format{
Column 1: individual ID.
Column 2: sire ID.
Column 3: dam ID.
Column 4: sex.
}
\details{
The pedigree is from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Xia Shen, Lars Ronnegard and Orjan Carlborg (2010): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. (14th QTLMAS, Poznan, Poland) SUBMITTED.
}
