\name{snpid}
\alias{snpid}
\docType{data}
\title{
Indices of The Selected SNPs of The Example Data
}
\description{
The indices of the selected SNPs of the example data.
}
\usage{data(snpid)}
\format{
a numeric vector containing SNP indices.
}
\details{
Every 6 SNPs were selected from the original dataset simulated for the 14th QTLMAS workshop, Poznan, Poland, 2010. See 'Source' and 'References'.
}
\source{
\url{http://jay.up.poznan.pl/qtlmas2010/dataset.html}
}
\references{
Shen X, Ronnegard L and Carlborg O (2011): Hierarchical likelihood opens a new way of estimating genetic values using dense marker maps. BMC Proceedings, 5(Suppl 3):S14.
}
